Endere til å hente ut getBigDecimal, da tablen inneholdet oracle NUMBER

+++ src/main/kotlin/no/nav/syfo/aksessering/db/ElektroniskAbonomentQueries.kt
@@ -24,5 +24,5 @@ fun DatabaseInterface.hentElektroniskAbonoment(herid: String): List<ElektroniskA
 
 fun ResultSet.toElektroniskAbonoment(): ElektroniskAbonoment =
         ElektroniskAbonoment(
-                getInt("partner_id")
+                getBigDecimal("partner_id").toInt()
         )
